Archives of the fourth module of the nanodegree machine learning engineer of the Udacity
docs:  I uploaded the files, still studying and preparing the internal
environment.

11-04-17 / 11:00

I'm still in the theoretical part of the course, but I'm already testing
the model. I tried to run the file but I could not find it and it was
because it was running in Python 3.5.2, and it should be running in 2.7.